style: my profile page (#608)
